Melhorando lógica do CRUD para Student e User.

+++ backend/src/main/java/br/edu/ifpb/backend/business/service/postgresql/StudentServicePostgresql.java
@@ -1,13 +1,13 @@
 package br.edu.ifpb.backend.business.service.postgresql;
 
 import br.edu.ifpb.backend.application.dto.request.StudentRequest;
+import br.edu.ifpb.backend.application.dto.request.UserRequest;
 import br.edu.ifpb.backend.business.entity.Student;
 import br.edu.ifpb.backend.business.entity.User;
 import br.edu.ifpb.backend.business.service.StudentService;
 import br.edu.ifpb.backend.repository.StudentRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -17,53 +17,64 @@ import java.util.List;
 public class StudentServicePostgresql implements StudentService {
     private StudentRepository studentRepository;
 
+    private UserServicePostgresql userServicePostgresql;
+
     public List<Student> index() {
         return studentRepository.findAll();
     }
 
     @Override
-    public void saveStudent(StudentRequest studentRequest) {
-        String encryptedPassword = new BCryptPasswordEncoder().encode(studentRequest.getPassword());
+    public boolean saveStudent(StudentRequest studentRequest) {
+        User userEntity = userServicePostgresql.registerUser(
+                new UserRequest(studentRequest.getUsername(), studentRequest.getPassword(), studentRequest.getEmail())
+        );
 
-        Student entity = new Student(1L, new User(1L, studentRequest.getUsername(),
-                    encryptedPassword,
-                    studentRequest.getEmail()),
+        Student studentEntity = new Student(
+                userEntity.getId(),
+                userEntity,
                 studentRequest.getName(),
                 studentRequest.getBirthDate(),
                 studentRequest.getCpf(),
-                studentRequest.getRg());
+                studentRequest.getRg()
+        );
 
-        studentRepository.save(entity);
+        studentRepository.save(studentEntity);
+
+        return true;
     }
 
     @Override
-    public void removeStudent(Long studentId) {
-        studentRepository.deleteById(studentId);
-    }
-
-    @Override
-    public void updateStudent(StudentRequest studentRequest, Long studentId) {
-        // String encryptedPassword = new BCryptPasswordEncoder().encode(studentRequest.getPassword());
-
-        Student studentById = null;
-
+    public boolean removeStudent(Long studentId) {
         try {
-            studentById = studentRepository.findById(studentId).orElseThrow(
-                    () -> new Exception("Student by ID " + studentId + " does not exist.")
-            );
-        } catch (Exception e) {
-            throw new RuntimeException(e);
+            if (studentRepository.existsById(studentId)) {
+                studentRepository.deleteById(studentId);
+                return true;
+            }
+        } catch (RuntimeException e) {
+            e.printStackTrace();
         }
+
+        return false;
+    }
+
+    @Override
+    public boolean updateStudent(StudentRequest studentRequest, Long studentId) {
+        Student studentById =  studentRepository.findById(studentId).orElseThrow(
+                () -> new RuntimeException("Student by ID " + studentId + " does not exist.")
+        );
 
         studentById.setBirthDate(studentRequest.getBirthDate());
         studentById.setCpf(studentRequest.getCpf());
         studentById.setName(studentRequest.getName());
         studentById.setRg(studentRequest.getRg());
 
-        studentById.getUser().setEmail(studentRequest.getEmail());
-        studentById.getUser().setUsername(studentRequest.getUsername());
-        // studentById.getUser().setPassword(encryptedPassword);
+        userServicePostgresql.updateUser(
+                new UserRequest(studentRequest.getUsername(), studentRequest.getPassword(), studentRequest.getEmail()),
+                studentById.getUser().getId()
+        );
 
         studentRepository.save(studentById);
+
+        return true;
     }
 }
